Invalid instruction defined for the ssl.sh script (apiman & apiman-gateway) #376

+++ elasticsearch/ssl.sh
@@ -1,11 +1,18 @@
 #!/bin/bash
 
-echo This script is called by maven to deploy the secret after the maven validate phase.
-echo Note that for this script to succeed you must oc login first
+#This script is called by maven to deploy the secret after the maven validate phase.
+#Note that for this script to succeed you must oc login first
+
+oc whoami > /dev/null
+if [ $? == 1 ]; then
+  echo "[ERROR] Not logged in, no 'elasticsearch-v1-keystore' secret is created"
+  exit 0;
+fi
 
 # deleting existing secret if there
-oc get secret elasticsearch-v1-keystore | grep 'Opaque' &> /dev/null
+oc get secret elasticsearch-v1-keystore &> /dev/null
 if [ $? == 0 ]; then
+  echo "[INFO] Delete existing secret/elasticsearch-v1-keystore"
   oc delete secret elasticsearch-v1-keystore
 fi
 # create the new secret
